refactor : #10 Product exception 디렉터리 오타 수정
