+++ com.devsquad10.product/src/main/java/com/devsquad10/product/application/exception/ProductNotFoundException.java
@@ -1,4 +1,4 @@
-package com.devsquad10.product.application.excption;
+package com.devsquad10.product.application.exception;
 
 public class ProductNotFoundException extends RuntimeException {
 	public ProductNotFoundException(String message) {
